+++ src/orchestrator/types.ts
@@ -1,6 +1,6 @@
-import { Address, Hex } from 'viem'
+import type { Address, Hex } from 'viem'
 import type { UserOperation } from 'viem/account-abstraction'
-import {
+import type {
   arbitrum,
   arbitrumSepolia,
   base,
@@ -34,6 +34,7 @@ type BundleStatus =
   | typeof BUNDLE_STATUS_PARTIALLY_COMPLETED
   | typeof BUNDLE_STATUS_COMPLETED
   | typeof BUNDLE_STATUS_FILLED
+  | typeof BUNDLE_STATUS_PRECONFIRMED
   | typeof BUNDLE_STATUS_FAILED
   | typeof BUNDLE_STATUS_UNKNOWN
 
@@ -120,14 +121,18 @@ interface WithoutOperation {
   targetExecutions?: never
 }
 
+type LockMode = 'HOOK' | 'COMPACT'
+
 interface MetaIntentBase {
-  targetChainId: number
-  tokenTransfers: TokenTransfer[]
   targetAccount: Address
+  targetChainId: number
+  targetGasUnits?: bigint
+  tokenTransfers: TokenTransfer[]
   accountAccessList?: {
     chainId: number
     tokenAddress: Address
   }[]
+  lockMode?: LockMode
   omniLock?: boolean
 }
 
@@ -275,6 +280,7 @@ const BUNDLE_STATUS_EXPIRED = 'EXPIRED'
 const BUNDLE_STATUS_PARTIALLY_COMPLETED = 'PARTIALLY_COMPLETED'
 const BUNDLE_STATUS_COMPLETED = 'COMPLETED'
 const BUNDLE_STATUS_FILLED = 'FILLED'
+const BUNDLE_STATUS_PRECONFIRMED = 'PRECONFIRMED'
 const BUNDLE_STATUS_UNKNOWN = 'UNKNOWN'
 
 export type {
@@ -309,5 +315,6 @@ export {
   BUNDLE_STATUS_PARTIALLY_COMPLETED,
   BUNDLE_STATUS_COMPLETED,
   BUNDLE_STATUS_FILLED,
+  BUNDLE_STATUS_PRECONFIRMED,
   BUNDLE_STATUS_UNKNOWN,
 }