+++ src/types.ts
@@ -1,5 +1,5 @@
-import { Account, Address, Chain, Hex } from 'viem'
-import { WebAuthnAccount } from 'viem/account-abstraction'
+import type { Account, Address, Chain, Hex } from 'viem'
+import type { WebAuthnAccount } from 'viem/account-abstraction'
 
 interface AccountProviderConfig {
   type: 'safe' | 'nexus'
@@ -134,6 +134,7 @@ type SignerSet = SessionSignerSet
 interface BaseTransaction {
   calls: Call[]
   tokenRequests: TokenRequest[]
+  gasLimit?: bigint
   signers?: SignerSet
 }
 
@@ -142,7 +143,7 @@ interface SameChainTransaction extends BaseTransaction {
 }
 
 interface CrossChainTransaction extends BaseTransaction {
-  sourceChain: Chain
+  sourceChain?: Chain
   targetChain: Chain
 }